requirements.txt
openai
typer
langchain
faiss-cpu
beautifulsoup4
requests
serpapi
python-dotenv
